More PR review comments
